requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-UUID==0.2
python-ldap==3.2.0
requests==2.21.0
Flask-QRcode==2.0.2
shortuuid==0.5.0
Flask-Login==0.4.1
Pillow==5.4.1
